\name{hmmpick-package}
\alias{hmmpick-package}
\alias{hmmpick}
\docType{package}
\title{
\packageTitle{hmmpick}
}
\description{
\packageDescription{hmmpick}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{hmmpick}
\packageIndices{hmmpick}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{hmmpick}

Maintainer: \packageMaintainer{hmmpick}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
