% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmmpick.R
\name{hmmpick}
\alias{hmmpick}
\title{Given a list of model files, chooses the best model}
\format{Returns a list sorted by ascending homogeneity cost.
        The models are used as element names, and the
        homogeneity cost are used as element values.}
\usage{
hmmpick(model_files)
}
\arguments{
\item{model_files}{List of ChromHMM model files to choose from}
}
\description{
Given a list of model files, chooses the best model
}
\details{
See \code{?homogeneity_cost} for the calculation of the
         homogeneity cost.
}
\examples{
model_files <- c(
  system.file("extdata", "model_roadmap15.txt",
              package="hmmpickr"),
  system.file("extdata", "model_roadmap18.txt",
              package="hmmpickr")
)
hmmpick(model_files)
}

