% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmmpick_cost_state.R
\name{hmmpick_cost_state}
\alias{hmmpick_cost_state}
\title{Cost of a state}
\usage{
hmmpick_cost_state(state_emissions)
}
\arguments{
\item{state_emissions}{list of state emissions}
}
\description{
Cost of a state
}
\details{
State cost calculated by
         \code{d_k = sum[h=1 to H]{min(1-E_hk, E_hk)}}
}

