% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmmpick_cost_state.R
\name{hmmpick_cost_states}
\alias{hmmpick_cost_states}
\title{Helper to calculate the cost of the different states
(used for development)}
\usage{
hmmpick_cost_states(emissions_probs, scale = 1, returntotal = TRUE)
}
\arguments{
\item{emissions_probs}{matrix (row=state; col=mark) containing the state emissions.}

\item{scale}{float/int or list (length should be equal to the number of rows in
\code{emissions_probs}) used to scale the states.
By default, no scaling is applied.}

\item{returntotal}{boolean; TRUE to return the average model cost or
FALSE to return the list of state costs.}
}
\description{
Helper to calculate the cost of the different states
(used for development)
}
\examples{
emissions_probs <- rbind(c(1, 0.5), c(1,0))
scale <- c(.95, .05)
hmmpick:::hmmpick_cost_states(emissions_probs, scale)
}

