% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homogeneity_cost.R
\name{homogeneity_cost}
\alias{homogeneity_cost}
\title{Calculate homogeneity cost}
\usage{
homogeneity_cost(x, is_model_file = TRUE)
}
\arguments{
\item{x}{Either the ChromHMM model file or
emission probabilities loaded from
\code{load_chromhmm_model(...)$emissions} or
\code{load_chromhmm_emissions(...)}}

\item{is_model_file}{Boolean; TRUE if \code{x} is the model file and
FALSE if \code{x} is emission probabilities.}
}
\description{
Calculate homogeneity cost
}
\details{
Calculated by \code{D = sum[k=1 to K]{d_k}^2/K} where
         \code{d_k = sum[h=1 to H]{min(1-E_hk, E_hk)}}.
}

