% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_chromhmm_emissions.R
\name{load_chromhmm_emissions}
\alias{load_chromhmm_emissions}
\title{Load ChromHMM emission probabilities}
\usage{
load_chromhmm_emissions(filename, ...)
}
\arguments{
\item{filename}{The ChromHMM model emissions file}

\item{...}{options for readr::read_tsv(...)}
}
\description{
Load ChromHMM emission probabilities
}

