% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_chromhmm_model.R
\name{load_chromhmm_model}
\alias{load_chromhmm_model}
\title{Load ChromHMM model}
\usage{
load_chromhmm_model(model_file, histone_order = NULL)
}
\arguments{
\item{model_file}{ChromHMM model file.
This file usually starts with "model_" and ends with ".txt"}

\item{histone_order}{Specify the histone order. Example
\code{c("H3K4me3", "H3K27ac", "H3K4me1", "H3K36me3",
        "H3K9me3", "H3K27me3")}}
}
\value{
Returns a list object.
\itemize{
 \item{"\code{num_states}"}{
   for the number of states in the model}
 \item{"\code{emissions}"}{
   for a dataframe with the emission probabilities}
 \item{"\code{transitions}"}{
   for a dataframe with the transition probabilities}
}
}
\description{
Load ChromHMM model
}
\examples{
model_file <-
  system.file("extdata", "model_roadmap15.txt",
              package="hmmpick")
model_file <-
  system.file("extdata", "model_roadmap18.txt",
              package="hmmpick")

load_chromhmm_model(model_file)

}

