% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_chromhmm_overlap_enrichments.R
\name{load_chromhmm_overlap_enrichments}
\alias{load_chromhmm_overlap_enrichments}
\title{Load ChromHMM overlap enrichment file}
\usage{
load_chromhmm_overlap_enrichments(enrichment_file, ...)
}
\arguments{
\item{enrichment_file}{ChormHMM model_*_emissions.txt file.
The first column of the TSV should be chromatin state.}

\item{...}{options for readr::read_tsv(...)}
}
\description{
Load ChromHMM overlap enrichment file
}
\details{
Replaces the name of the first column with "state" and
         filters out the "Base" state.
}

