% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmmpick.R
\name{hmmpick}
\alias{hmmpick}
\title{Pick the ChromHMM model that is most well-defined}
\format{Returns a list sorted by ascending homogeneity cost.
        The models in \code{model_files} are used as the element names, and
        the model homogeneity cost are used as the element values.}
\usage{
hmmpick(model_files, full_path = FALSE)
}
\arguments{
\item{model_files}{A list of ChromHMM models}

\item{full_path}{Boolean; TRUE to return the full model path in the output,
or FALSE to return only the model basename.}
}
\description{
Pick the ChromHMM model that is most well-defined
}
\examples{
model_files <- c(
  system.file("extdata", "model_roadmap15.txt",
              package="hmmpickr"),
  system.file("extdata", "model_roadmap18.txt",
              package="hmmpickr")
)
hmmpick(model_files)
}
\seealso{
\code{\link{homogeneity_cost}} for the calculation of the
         homogeneity cost.
}

