% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_chromhmm_emissions.R
\name{load_chromhmm_emissions}
\alias{load_chromhmm_emissions}
\title{Load ChromHMM enrichment file}
\usage{
load_chromhmm_emissions(enrichment_file)
}
\arguments{
\item{enrichment_file}{ChormHMM model_*_emissions.txt file.
The first column of the TSV should be chromatin state.}
}
\description{
Load ChromHMM enrichment file
}
\details{
Replaces the name of the first column with "state" and
         filters out the "Base" state.
}

